+++ src/dominio/Loan.java
@@ -9,9 +9,9 @@ public class Loan {
 	private String dni;
 	private int accountNumber;
 	private Date loanDate;
-	private Float amountReqByCustomer;
-	private Float amountInt;
-	private Float monthlyFee;
+	private Double amountReqByCustomer;
+	private Double amountInt;
+	private Double monthlyFee;
 	private Date paymentDeadline;
 	private int amountOfFees;
 	private LoanState loanState;
@@ -20,9 +20,10 @@ public class Loan {
 	public Loan() {
 
 	}
+
 	
-	public Loan(int id, String dni, int accountNumber, Date loanDate, Float amountReqByCustomer, Float amountInt,
-			Float monthlyFee, Date paymentDeadline, int amountOfFees, LoanState loanState) {
+	public Loan(int id, String dni, int accountNumber, Date loanDate, Double amountReqByCustomer, Double amountInt,
+			Double monthlyFee, Date paymentDeadline, int amountOfFees, LoanState loanState) {
 		super();
 		this.id = id;
 		this.dni = dni;
@@ -35,86 +36,107 @@ public class Loan {
 		this.amountOfFees = amountOfFees;
 		this.loanState = loanState;
 	}
+	
 
 	public int getId() {
 		return id;
 	}
 
+
 	public void setId(int id) {
 		this.id = id;
 	}
+
 
 	public String getDni() {
 		return dni;
 	}
 
+
 	public void setDni(String dni) {
 		this.dni = dni;
 	}
+
 
 	public int getAccountNumber() {
 		return accountNumber;
 	}
 
+
 	public void setAccountNumber(int accountNumber) {
 		this.accountNumber = accountNumber;
 	}
+
 
 	public Date getLoanDate() {
 		return loanDate;
 	}
 
+
 	public void setLoanDate(Date loanDate) {
 		this.loanDate = loanDate;
 	}
 
-	public Float getAmountReqByCustomer() {
+
+	public Double getAmountReqByCustomer() {
 		return amountReqByCustomer;
 	}
 
-	public void setAmountReqByCustomer(Float amountReqByCustomer) {
+
+	public void setAmountReqByCustomer(Double amountReqByCustomer) {
 		this.amountReqByCustomer = amountReqByCustomer;
 	}
 
-	public Float getAmountInt() {
+
+	public Double getAmountInt() {
 		return amountInt;
 	}
 
-	public void setAmountInt(Float amountInt) {
+
+	public void setAmountInt(Double amountInt) {
 		this.amountInt = amountInt;
 	}
 
-	public Float getMonthlyFee() {
+
+	public Double getMonthlyFee() {
 		return monthlyFee;
 	}
 
-	public void setMonthlyFee(Float monthlyFee) {
+
+	public void setMonthlyFee(Double monthlyFee) {
 		this.monthlyFee = monthlyFee;
 	}
+
 
 	public Date getPaymentDeadline() {
 		return paymentDeadline;
 	}
 
+
 	public void setPaymentDeadline(Date paymentDeadline) {
 		this.paymentDeadline = paymentDeadline;
 	}
+
 
 	public int getAmountOfFees() {
 		return amountOfFees;
 	}
 
+
 	public void setAmountOfFees(int amountOfFees) {
 		this.amountOfFees = amountOfFees;
 	}
+
 
 	public LoanState getLoanState() {
 		return loanState;
 	}
 
+
 	public void setLoanState(LoanState loanState) {
 		this.loanState = loanState;
 	}
+
 
 	@Override
 	public String toString() {
@@ -123,7 +145,4 @@ public class Loan {
 				+ monthlyFee + ", paymentDeadline=" + paymentDeadline + ", amountOfFees=" + amountOfFees
 				+ ", loanState=" + loanState + "]";
 	}
-	
-	
-	
 }
